+++ psi4/src/psi4/libfock/points.cc
@@ -699,7 +699,7 @@ void BasisFunctions::compute_functions(std::shared_ptr<BlockOPoints> block) {
         center[2] = v[2];
 
         // Make new pointers, gg computes along rows so we need to skip down `nval` rows.
-        size_t row_shift = nvals * npoints;
+        size_t row_shift = static_cast<size_t> nvals * npoints;
         double* phi_start = tmpp + row_shift;
 
         // Copmute collocation
